+++ Cargo.toml
@@ -39,6 +39,7 @@ tracing-subscriber = {version = "0.3"}
 tracing-actix-web = "0.7"
 derive_more = "0.99"
 tokio-util = "0.7"
+async-trait = "0.1"
