+++ src/mod_ex_flow_utils/mod.rs
@@ -1,3 +1,4 @@
 pub mod utils_ex_flow;
 pub mod entities;
 pub mod uri;
+pub mod datatype_ex_flow;
